requirements/base.txt
pytz==2020.1
python-slugify==4.0.0
Pillow
argon2-cffi==20.1.0
redis
celery
django-celery-beat
flower
requests
django
django-environ
django-model-utils
django-crispy-forms==1.9.1
django-redis==4.12.1
alabaster==0.7.12
attrs==19.3.0
cfgv==3.1.0
cookiecutter==1.7.0
decorator==4.4.2
defusedxml==0.6.0
django-countries
django-widget-tweaks==1.4.8
djangorestframework
docutils==0.16
entrypoints==0.3
future==0.18.2
imagesize==1.2.0
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.1
psycopg2-binary==2.8.4
py==1.10.0
python3-openid==3.1.0
regex
six==1.14.0
urllib3==1.25.8
virtualenv==20.0.7
zipp==3.0.0
django-cors-headers==3.5.0
django-ckeditor
django-qr-code==1.3.1
django-otp
qrcode
django-taggit
base58
tldextract
pycryptodome==3.9.7
django_compressor
coinbase
